Move lists of servlets to sub-modules.
